pyproject.toml
[tool.poetry]
name = "kurs-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kurs_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.12"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.3.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.0"
django-celery-beat = "^2.5.0"
telebot = "^0.0.5"
redis = "^5.0.1"
celery = "^5.3.4"
eventlet = "^0.33.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
